QuantReg.compute_inf! — Method

compute_inf!(model)

In-place version of [compute_inf!(model)]

QuantReg.compute_inf — Method

compute_inf(model::QuantRegModel)

Compute inference for model as specified in model.inf.

QuantReg.rq — Method

rq(formula::FormulaTerm, data::DataFrame; kargs)

Generate, fit, and compute inference for the specified quantile regression model.

Keyword Arguments

StatsBase.coeftable — Method

coeftable(model::QuantRegModel)

Generate a coefficient table from a QuantRegModel.

QuantReg.QuantRegModel — Type

QuantRegModel(formula::FormulaTerm, data::DataFrame; τ::Number=0.5,
              fitmethod::String="br",invers::Union{Nothing, Bool}=nothing,
              α::Number=0.05, hs::Bool=true, iid::Union{Nothing, Bool}=nothing,
              interpolate::Bool=true, tcrit::Bool=true)

Quantile Regression model at the τth quantile fitting data according to formula.

In any call, formula and data must be set. Additional parameters and their defaults are as specified below:

τ: the quantile to fit the model at; default is 0.5, i.e. median regression

fitmethod: the method to fit the model with; avaliable options are:

"br": fit using Barrodlae-Roberts simplex (default method); see [fitbr!] for

details

"fn": Fit using Frisch-Newton algorithm; see [fitfn!`] for details

"gurobi": Fit using Gurobi (must have Gurobi installed); see [fitgurobi!``]

invers: if true, compute confidence intervals by inverting a rank test (otherwise use an

asymptotic esimtate of the covariance matrix); default setting is datasets with 1000 or fewer observations and false for larger datasets

α: size of test for computing inference; default setting is 0.05

hs: if true, use Hall Sheather bandwidth when computing sparsity esimtates

(otherwise, use Bofinger bandwidth); default is true

iid: if true, assume model errors are iid (otherwise, assume that the conditional

quantile function is locally (in tau) linear (in x)); default is true if using rank test inversion and false if using an asymptotic estimate of the covariance matrix

interpolate: if true, interpolate the confidence intervals produced by rank test

inversion inference (otherwise, print values just above and below); default is true

tcrit: if true, use a Student's t distribution for calculating critical points

(otherwise use a normal distribution); default is true

QuantReg.QuantRegModel — Method

QuantRegModel(model::QuantRegModel; τ::Union{Nothing, Number}=nothing,
              fitmethod::Union{Nothing, String}=nothing,
              invers::Union{Nothing, Bool}=nothing, α::Union{Nothing, Number}=nothing,
              hs::Union{Nothing, Bool}=nothing, iid::Union{Nothing, Bool}=nothing,
              interpolate::Union{Nothing, Bool}=nothing,
              tcrit::Union{Nothing, Bool}=nothing)

Construct a new QuantileRegression model by changing a parameter in an existing model.

If τ or method are unchanged, then the model fit is retained from the passed model but the model inference type is not.

Base.append! — Method

append!(X::QuantRegModels, model::QuantRegModel)

Add model to X in-place; throws an error if X already contains a model with the same τ value as model.

Base.copy — Method

copy(fit::QuantRegFit)

Create a deep copy of fit.

Base.copy — Method

copy(inf::QuantRegInf)

Create a deep copy of inf.

Base.copy — Method

copy(model::QuantRegModel)

Create a deep copy of model (excluding data and formula).

Base.getindex — Method

getindex(X::QuantRegModels, τ::Number)

Returns the model in X fit at the τth quantile.

Base.show — Method

show(io::IO, models:QuantRegModels)

Display each model in models.

Base.show — Method

show(io::IO, model:QuantRegModel)

Display quantreg model.

QuantReg.compute_bandwidth — Method

calcbandwidth(τ::Number, n::Integer, α::Number; hs=false)

Calculate the τ bandwidth for sparsity estimation when calculating inference under the assumption that the conditional quantile function is locally (in tau) linear (in x).

Arguments

τ: quantile

n: sample size

α: alpha level for intended confidence interval

hs: if true, use Hall Sheather bandwidth; otherwise, use Bofinger bandwidth

QuantReg.compute_nid_qn_invers — Method

compute_nid_qn_invers(i::Integer, data::DataFrame, regressors::Array{<:AbstractTerm},
                  weights::Array{<:Number})

Computes residuals variances from the projection of each column of X on remaining columns for rank test inversion inference under the n.i.d. assumption.

This function should be of little interest to end users as it is primarily a helper function for computing inference with a rank test inversion and with n.i.d. errors.

Arguments

i: enumeration of column to project

data: data used to fit the model

regressors: list of regressors from the model

weights: weight to use in calculating projection as a linear regression

QuantReg.compute_σ_iid_asy — Method

compute_σ_iid_asy(model::QuantRegModel)

Compute standard errors for a model using an estimate of the asymptotic covariance matrix under the assumtion that errors are iid.

QuantReg.compute_σ_nid_asy — Method

compute_σ_nid_asy(model::QuantRegModel)

Compute standard errors for a model using an estimate of the asymptotic covariance matrix under the assumption that the conditional quantile function is locally (in tau) linear (in x).

QuantReg.fit — Method

fit(model::QuantRegModel)

Fit model according to model.fit.method.

QuantReg.fitbr! — Method

fitbr!(model::QuantRegModel; ci::Bool=false),

Fit model using the Barrodale-Roberts method.

If ci is false, model.fit is updated in place to reflect the fit produced by running the Barrodale-Roberts simplex, and confidence intervals are not computed. Otherwise, confidence intervals are computed, and model.inf is updated in place to reflect the confidence invervals produced by this method, but model.fit is not updated.

This fitting method leverages public domain FORTRAN code written by Roger Koenker for the R quantreg package.

QuantReg.fitbrfortran — Method

fitbrfortran(n::Integer, k::Integer, X::Matrix{Number}, y::Vector{Number},
             τ::Number, nsol::Integer, ndsol::Integer, qn::Vector{Number},
             cutoff::Number, lci1::Bool)

Wraps call to the public domain Barrodale-Roberts simplex FORTRAN routine.

Arguments

n: number of observations

k: number of parameters

X: x matrix

y: y vector

τ: the desired quantile

nsol: an estimated (row) dimension of the primal solution array

ndsol1: an estimated (row) dimension of the dual solution array

qn: residuals variances from the projection of each column of X on remaining columns

cutoff: the critical point for testing

lci1: whether to calculate CI

QuantReg.fitfn! — Method

fitfn(model::QuantRegModel)

Fit model using the Frish-Newton algorithm.

Fitting with this method does not produce solutions to the dual problem.

This fitting method leverages public domain FORTRAN code written by Roger Koenker for the R quantreg package. In the quantreg packaage, this is equivalent to the fn and fnb methods.

QuantReg.fitgurobi! — Method

fitgurobi(model::QuantRegModel)

Fit model using Gurobi via Julia's JuMP library.

This algorithm has been tailored specifically to work with Julia (hence the use of direct) mode. Some time was spent researching the use of open-source solvers such as GLPK, but they proved to be too slow. Extending this method to work with other solvers, such as CPLEX could be accomplished by switching the optimizer and creating the model in automatic mode, but this may leave some small efficiency gains on the table.

QuantReg.hastau — Method

getindex(X::QuantRegModels, τ::Number)

Check if X contains a model at the τth percentile.

QuantReg.init_ci_invers — Method

init_inf_invers(model::QuantRegModel)

Initialize necessary values for calculating confidence intervals for model with a rank test inversion.

This function should be of little interest to end users as it is solely used as a subroutine when calling [fitbr(model; ci=true)].

QuantReg.taus — Method

taus(X::QuantRegModels)

Retrive all the τ values for models stored in X.

QuantReg.write_ci! — Method

write_ci(model::QuantRegModel, ci::Array{Number, 2})

Write confidence intervals to model.inf.

ci should be a set of confidence interval matrix produced by a call to [fitbr(model; ci=true)].

StatsBase.fit! — Method

fit!(model::QuantRegModel)

In-place version of [fit(model)].

QuantReg.QuantRegFit — Type

QuantRegFit(computed::Bool, coef::Union{Vector, Nothing}, resid::Union{Vector, Nothing},
            dual::Union{Array, Nothing}, yhat::Union{Vector, Nothing})

Stores results of fitting a quantile regression model.

coef, resid, dual, and yhat should be set to nothing until a model is fit.

QuantReg.QuantRegInf — Type

QuantRegInf(computed::Bool, lowerci::Union{Nothing, Array{Number}},
            upperci::Union{Nothing, Array{Number}}, σ::Union{Nothing, Array{Number}},
            teststat::Union{Nothing, Array{Number}}, p::Union{Nothing, Array{Number}})

Contains specifcations for and results of computing inference for a quantile regression model.

lowerci, upperci, σ, teststat should be set to nothing until inference is computed.

QuantReg.QuantRegModels — Type

QuantRegModels()

Wrapper containing multiple QuantRegModel at different quantiles indexed by quantile.

This method partically implements some behaviors of a dictionary (for easy indexing) and some behaviors of an array (for easy, consistent appends). This type is not intended to be directly created by most end users.

For example, if a models::QuantRegModels contained models with τ=0.25, 0.5 and 0.75, these models could be accessed as models[0.25], models[0.5], and models[0.75] respectively.